refactor(system): separate upgrade to types
